% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSpace.R
\name{CladeStats}
\alias{CladeStats}
\title{CladeStats}
\usage{
CladeStats(data.obj, rightSVD.cells, heights, cores)
}
\arguments{
\item{data.obj}{Seurat object containing scATACseq data.}

\item{rightSVD.cells}{List containing right singular values and clustering information.}

\item{heights}{Numeric vector of heights for dendrogram cuts.}

\item{cores}{Integer, number of CPU cores for parallel processing.}
}
\value{
A list containing cluster statistics calculated for each height.
}
\description{
CladeStats
}
\examples{
\dontrun{
heights <- seq(0.01, 1, by=0.01)
clade.stats <- CladeStats(data.obj, RightSVD_dist(data.obj), heights, cores)
print(clade.stats)
}
}
