% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSpace.R
\name{CreateChromAssay}
\alias{CreateChromAssay}
\title{CreateChromAssay}
\usage{
CreateChromAssay(data.list, genome = "mm10")
}
\arguments{
\item{data.list}{A list of lists for each replicate. The sublists need to include list components "feature.matrix" and "fragmentObject".}

\item{genome}{UCSC name of the used genome.}
}
\value{
Returns a list
}
\description{
Creates a \code{\link[Signac]{ChromatinAssay}} object for every replicate on a list.
}
\details{
Part of the wrapper function \code{\link{ProcessReplicates}}. Based on the Signac function \code{\link[Signac]{CreateChromatinAssay}}.
}
\examples{
\dontrun{
data.list <- list(list(feature.matrix=matrix(rpois(100, lambda=5), nrow=10),\
 fragmentObject="fragment_path"))
chrom_assay <- CreateChromAssay(data.list, genome="mm10")
print(chrom_assay)
}
}
