% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSpace.R
\name{CreateFeatureMatrix}
\alias{CreateFeatureMatrix}
\title{CreateFeatureMatrix}
\usage{
CreateFeatureMatrix(data.list, binsize = 5000, process_n = 5000)
}
\arguments{
\item{data.list}{A list of lists for each replicate. The sublists need to have list components "fragmentObject" and "genome.lengths".}

\item{binsize}{Size of the genome bins to use}

\item{process_n}{Number of regions to load into memory at a time, per thread. Processing more regions at once can be faster but uses more memory.}
}
\value{
Returns a list
}
\description{
Constructs a bin x cell matrix for each replicate on a list.
}
\details{
Part of the wrapper function \code{\link{ProcessReplicates}}. Based on the Signac function \code{\link[Signac]{GenomeBinMatrix}}.
}
\examples{
\dontrun{
data.list <- list(list(fragmentObject="fragment_path", genome.lengths=c(chr1=100000)))
feature_matrix <- CreateFeatureMatrix(data.list, binsize=5000, process_n=5000)
print(feature_matrix)
}
}
