% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSpace.R
\name{CreateFragmentObjects}
\alias{CreateFragmentObjects}
\title{CreateFragmentObjects}
\usage{
CreateFragmentObjects(data.list, validate.fragments = FALSE)
}
\arguments{
\item{data.list}{A list with list components "filtered.fragment.name" (a path to a fragment file) and "true.cell.barcodes" (the barcodes of the fragments).}

\item{validate.fragments}{Check that expected cells are present in the fragment file.}
}
\value{
Returns a list
}
\description{
Creates a \code{Fragment} object for every replicate on a list.
}
\details{
Part of the wrapper function \code{\link{ProcessReplicates}}. Based on the Signac function \code{\link[Signac]{CreateFragmentObject}}.
}
\examples{
\dontrun{
data.list <- list(
  filtered.fragment.name = "path/to/fragment.tsv.gz",
  true.cell.barcodes = c("cell1", "cell2")
)
fragment_objects <- CreateFragmentObjects(data.list, validate.fragments = TRUE)
print(fragment_objects)
}
}
