% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helpers.R
\name{FeatureSpace-config}
\alias{FeatureSpace-config}
\title{Global Configuration}
\arguments{
\item{cores}{Number of CPU cores to use for parallel processing}

\item{data_path}{Path to input data directory}

\item{macs2_path}{Path to MACS2 tool for peak calling}
}
\description{
Configuration settings for FeatureSpace package
}
