% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSpace.R
\name{FilterFragments}
\alias{FilterFragments}
\title{FilterFragments}
\usage{
FilterFragments(data.list, force = FALSE, buffer_length = 256L)
}
\arguments{
\item{data.list}{A list of replicates each with a fragments file path and a list component "true.cell.barcodes".}

\item{force}{If TRUE recreates an output file even if it already exists.}

\item{buffer_length}{Size of buffer to be read from the fragment file. This must be longer than the longest line in the file.}
}
\value{
Returns a list with a path to a fragment file.
}
\description{
Takes a list of replicates each with a fragments file path and a list component "true.cell.barcodes". Removes fragments from the fragment file if their barcodes are not included in "true.cell.barcodes". The output is a gzip-compressed and indexed file.
}
\details{
Part of the wrapper function \code{\link{ProcessReplicates}}. Based on the Signac function \code{\link[Signac]{FilterCells}}.
}
\examples{
\dontrun{
filtered_fragments <- FilterFragments(fragments)
print(filtered_fragments)
}
}
