% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSpace.R
\name{FindDataFiles}
\alias{FindDataFiles}
\title{FindDataFiles}
\usage{
FindDataFiles(data.paths, sample.names, barcode.file = "*singlecell.csv")
}
\arguments{
\item{data.paths}{Path to the folder with scATAC data including fragments and barcodes.}

\item{sample.names}{Names for each replicate.}

\item{barcode.file}{The file name or pattern of the file with cell barcodes.}
}
\value{
Returns a list
}
\description{
Finds the relevant scATAC data and barcode files and writes their paths in a list. This is done for each replicate, so the result is a list of lists.
}
\details{
Part of the wrapper function \code{\link{ProcessReplicates}}.
}
\examples{
\dontrun{
data.paths <- c("path/to/sample1", "path/to/sample2")
sample.names <- c("sample1", "sample2")
barcode.file <- "singlecell.csv"
data_files <- FindDataFiles(data.paths, sample.names, barcode.file)
print(data_files)
}
}
