% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSpace.R
\name{FrequencyFilter}
\alias{FrequencyFilter}
\title{FrequencyFilter}
\usage{
FrequencyFilter(
  data.object,
  min.cells = 0.025,
  max.cells = 0.975,
  min.features = 0.025,
  max.features = 0.975
)
}
\arguments{
\item{data.object}{A \code{Seurat} object including a "cladepeaks_" \code{\link[Signac]{ChromatinAssay}}.}

\item{min.cells}{A threshold for the smallest amount of cells per feature. Given as a percentage of all cells in "data.object".}

\item{max.cells}{A threshold for the largest amount of cells per feature. Given as a percentage of all cells in "data.object".}

\item{min.features}{A threshold for the smallest amount of features per cell. Given as a percentage of all features in "data.object".}

\item{max.features}{A threshold for the largest amount of features per cell. Given as a percentage of all features in "data.object".}
}
\value{
Returns a \code{Seurat} object.
}
\description{
Removes cells with too few or too many features and features found in too few or too many cells.
}
\examples{
\dontrun{
data.obj <- CreateSeuratObject(matrix(rnorm(100), nrow=10))
filtered_obj <- FrequencyFilter(data.obj, min.cells=0.025, max.cells=0.975,\
 min.features=0.025, max.features=0.975)
print(filtered_obj)
}
}
