% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSpace.R
\name{MergeReplicates}
\alias{MergeReplicates}
\title{MergeReplicates}
\usage{
MergeReplicates(data.list, sample.names)
}
\arguments{
\item{data.list}{A list of lists for each replicate. The sublists need to have a list component with the name "seurat.obj".}

\item{sample.names}{Names for each replicate.}
}
\value{
Returns a list
}
\description{
Merges the \code{Seurat} objects (one per replicate) on a list and adds the result as a new list component.
}
\details{
Part of the wrapper function \code{\link{ProcessReplicates}}.
}
\examples{
\dontrun{
data.list <- list(
  list(seurat.obj = CreateSeuratObject(matrix(rnorm(100), nrow=10))),
  list(seurat.obj = CreateSeuratObject(matrix(rnorm(100), nrow=10)))
)
sample.names <- c("replicate1", "replicate2")
merged_data <- MergeReplicates(data.list, sample.names)
print(merged_data)
}
}
