% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSpace.R
\name{PlotCladeStats}
\alias{PlotCladeStats}
\title{PlotCladeStats}
\usage{
PlotCladeStats(clade.stats)
}
\arguments{
\item{clade.stats}{A list containing clustering statistics including Dunn index and silhouette width.}
}
\value{
A ggplot2 object visualizing clustering statistics across different cut heights.
}
\description{
PlotCladeStats
}
\examples{
\dontrun{
heights <- seq(0.01, 1, by=0.01)
clade.stats <- CladeStats(E14VMB, RightSVD_dist(E14VMB), heights, cores = 2)
PlotCladeStats(clade.stats)
}
}
