% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSpace.R
\name{PlotCladeTree}
\alias{PlotCladeTree}
\title{PlotCladeTree}
\usage{
PlotCladeTree(hcd, cut.height)
}
\arguments{
\item{hcd}{Out put from RightSVD_dist()}

\item{cut.height}{The height under which all branches are combined into a cell clade.}
}
\value{
Draws a plot.
}
\description{
Draws a tree of cell clades based on cut height.
}
\examples{
\dontrun{
hclust.obj <- hclust(dist(matrix(rnorm(100), nrow=10)))
PlotCladeTree(hclust.obj)
}
}
