% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSpace.R
\name{ProcessReplicates}
\alias{ProcessReplicates}
\title{ProcessReplicates}
\usage{
ProcessReplicates(
  data.paths,
  sample.names,
  barcode.file = "*singlecell.csv",
  is.cell.name = "is__cell",
  force = FALSE,
  validate.fragments = FALSE,
  genome = "mm10",
  genome.obj = BSgenome.Mmusculus.UCSC.mm10,
  used.chromosomes = c("chr1", "chr2", "chr3", "chr4", "chr5", "chr6", "chr7", "chr8",
    "chr9", "chr10", "chr11", "chr12", "chr13", "chr14", "chr15", "chr16", "chr17",
    "chr18", "chr19"),
  binsize = 5000,
  process_n = 5000,
  buffer_length = 256L
)
}
\arguments{
\item{data.paths}{Path to the folder with scATAC data including fragments and barcodes.}

\item{sample.names}{Names for each replicate.}

\item{barcode.file}{The file name or pattern of the file with cell barcodes.}

\item{is.cell.name}{A column in the barcode file that tells if the barcode in question is a cell barcode (1) or not (0).}

\item{force}{If TRUE recreates an output file for filtered fragments even if it already exists.}

\item{validate.fragments}{Check that expected cells are present in the fragment file.}

\item{genome}{UCSC name of the used genome.}

\item{genome.obj}{An object containing sequence information.}

\item{used.chromosomes}{The chromosomes used in the analysis.}

\item{binsize}{Size of the genome bins to use}

\item{process_n}{Number of regions to load into memory at a time, per thread. Processing more regions at once can be faster but uses more memory.}

\item{buffer_length}{Size of buffer to be read from the fragment file. This must be longer than the longest line in the file.}
}
\value{
Returns a list
}
\description{
A wrapper function that creates a list of processed genomic and scATAC information for each replicate, thus the output is a list of lists.
}
\details{
Uses FeatureSpace functions \code{\link{FindDataFiles}}, \code{\link{ReadBarcodeMetadata}}, \code{\link{FilterFragments}}, \code{\link{CreateFragmentObjects}}, \code{\link{SetUsedChromosomes}}, \code{\link{CreateFeatureMatrix}}, \code{\link{CreateChromAssay}}, \code{\link{CreateSeuratObj}} and \code{\link{MergeReplicates}}.
}
\examples{
\dontrun{
p <- paste(data_path, "e14vmb1_data/scATAC/E14VMB_0", sep="")
E14VMB <- ProcessReplicates(p, sample.names=c("_0"),\
 is.cell.name="is__cell_barcode", buffer_length=512L)
print(E14VMB)
}
}
