% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSpace.R
\name{ReadBarcodeMetadata}
\alias{ReadBarcodeMetadata}
\title{ReadBarcodeMetadata}
\usage{
ReadBarcodeMetadata(data.list, is.cell.name = "is__cell_barcode")
}
\arguments{
\item{data.list}{A list including a path to a file with barcode information for each replicate.}

\item{is.cell.name}{A column in the barcode file that tells if the barcode in question is a cell barcode (1) or not (0).}
}
\value{
Returns a list
}
\description{
Adds list components to a list including a path to a barcode file. The components are the barcode file as a tibble ("barcode.metadata") and a character vector including only cell barcodes ("true.cell.barcodes").
}
\details{
Part of the wrapper function \code{\link{ProcessReplicates}}.
}
\examples{
\dontrun{
barcode_metadata <- ReadBarcodeMetadata("path/to/barcode_metadata.csv")
}
}
