% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSpace.R
\name{RightSVD_dist}
\alias{RightSVD_dist}
\title{RightSVD_dist}
\usage{
RightSVD_dist(data.object, drop.components = 1)
}
\arguments{
\item{data.object}{A Seurat object containing chromatin accessibility data.}

\item{drop.components}{Numeric or integer vector specifying the SVD components to exclude from the analysis. Default is 1.}
}
\value{
A list containing:
\item{hclust.cells}{Hierarchical clustering object of cells.}
\item{cell.dist}{A distance matrix of cells computed using cosine distance.}
}
\description{
RightSVD_dist
}
\examples{
\dontrun{
rightSVD <- RightSVD_dist(E14VMB)
print(rightSVD$hclust.cells)
print(rightSVD$cell.dist)
}
}
