% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSpace.R
\name{SetUsedChromosomes}
\alias{SetUsedChromosomes}
\title{SetUsedChromosomes}
\usage{
SetUsedChromosomes(
  data.list,
  genome.obj = BSgenome.Mmusculus.UCSC.mm10,
  used.chromosomes = c("chr1", "chr2", "chr3", "chr4", "chr5", "chr6", "chr7", "chr8",
    "chr9", "chr10", "chr11", "chr12", "chr13", "chr14", "chr15", "chr16", "chr17",
    "chr18", "chr19")
)
}
\arguments{
\item{data.list}{A list of 1 or more list components.}

\item{genome.obj}{An object containing sequence information.}

\item{used.chromosomes}{The chromosomes used in the analysis.}
}
\value{
Returns a list
}
\description{
Creates a list component (for every replicate) with chromosome sequence lengths for every selected chromosome.
}
\details{
Part of the wrapper function \code{\link{ProcessReplicates}}.
}
\examples{
\dontrun{
SetUsedChromosomes(c("chr1", "chr2", "chr3"))
}
}
